moves methods to thought repository
