+++ thoughts/models/thought.py
@@ -2,14 +2,14 @@ from django.db import models
 from django.contrib.auth.models import User
 
 class Thought(models.Model):
-  text = models.TextField()
-  date_created = models.DateTimeField(auto_now_add=True)
-  date_updated = models.DateTimeField(auto_now=True)
-  is_locked = models.BooleanField(default=True)
-  user = models.ForeignKey(User)
+    text = models.TextField()
+    date_created = models.DateTimeField(auto_now_add=True)
+    date_updated = models.DateTimeField(auto_now=True)
+    is_locked = models.BooleanField(default=True)
+    user = models.ForeignKey(User)
 
-  def __unicode__(self):
-    return """Thought ID: {0}
+    def __unicode__(self):
+        return """Thought ID: {0}
               Created: {1}
               Updated: {2}
               Locked?: {3}
@@ -18,34 +18,15 @@ class Thought(models.Model):
                                              self.date_updated,
                                              self.is_locked,
                                              self.user
-                                            )
+                                             )
 
-  @classmethod
-  def locked_thought_count(cls):
-    return Thought.objects.filter(is_locked=True).count()
+    def save(self, pool_size=0, *args, **kwargs):
+        if Thought._ready_to_unlock(pool_size):
+            self.is_locked = False
+            super (Thought, self).save(*args, **kwargs)
+        else:
+            super(Thought, self).save(*args, **kwargs)
 
-  @classmethod
-  def unlocked_thought_count(cls):
-    return Thought.objects.filter(is_locked=False).count()
-
-  @classmethod
-  def next_locked_thought(cls):
-    return Thought.objects.filter(is_locked=True).order_by('date_created').first()
-
-  @classmethod
-  def unlock_thoughts(cls, unlocked_pool_size):
-    if Thought.locked_thought_count() > 0:
-      thought_keys = Thought.objects.filter(is_locked=True).order_by('date_created').values('pk')[:unlocked_pool_size]
-      return Thought.objects.filter(pk__in=thought_keys).update(is_locked=False)
-    return 0
-
-  @classmethod
-  def ready_to_unlock(cls, pool_size):
-    return True if pool_size > 0 else False
-
-  def save(self, pool_size=0, *args, **kwargs):
-    if Thought.ready_to_unlock(pool_size):
-      self.is_locked = False
-      super(Thought, self).save(*args, **kwargs)
-    else:
-      super(Thought, self).save(*args, **kwargs)
+    @classmethod
+    def _ready_to_unlock(cls, pool_size):
+        return True if pool_size > 0 else False
